Am corectat greselile din cod
