+++ Laborator6/src/Constants/Constants.java
@@ -1,0 +1,15 @@
+package Constants;
+
+import SERVICE.Pisica;
+import model1.Animal;
+
+import java.util.List;
+import java.util.Objects;
+
+public class constante {
+    public final String LISTA_ANIMALE_GOALA = "Niciun animal in lista";
+
+    public static final String LISTA_VACCINURI_GOALA = "Animalul nu are niciun vaccin";
+
+
+}
