+++ Laborator6/src/Exceptions/ExceptieListaAnimaleGoala.java
@@ -1,0 +1,8 @@
+package Exceptions;
+
+public class Exceptie2 extends RuntimeException{
+
+    public Exceptie2(String message){
+        super(message);
+    }
+}
