+++ Laborator6/src/Exceptions/ExceptieListaVaccinuriGoala.java
@@ -1,0 +1,9 @@
+package Exceptions;
+
+public class ExceptieListaVaccinuriGoala extends RuntimeException{
+
+    public ExceptieListaVaccinuriGoala(String message){
+        super(message);
+    }
+
+}
