+++ Laborator6/src/Exceptions/ExceptieVaccinIntarziat.java
@@ -1,0 +1,8 @@
+package model1.exceptii;
+
+public class VaccinIntarziatexceptie extends RuntimeException{
+
+    public VaccinIntarziatexceptie(String m){
+        super(m);
+    }
+}
