+++ Laborator6/src/Models/Animal.java
@@ -1,0 +1,8 @@
+package model1;
+
+public interface Animal {
+
+    public FisaMedicala getFisaMedicala();
+
+    public String vorbeste();
+}
