--- Laborator6/src/SERVICE/Pisica.java
@@ -1,30 +1,0 @@
-package SERVICE;
-
-public class Pisica {
-
-    public String nume;
-    private int Varsta;
-
-
-    public Pisica(String nume){
-        this.nume = nume;
-    }
-
-    public Pisica(int v){
-        this.Varsta = v;
-    }
-
-    public Pisica(){
-
-    }
-
-    private void setNume(String n){
-        this.nume = nume;
-        this.Varsta = 1;
-    }
-
-    public String vorbeste(){
-        return "Pisica " + this.nume + " miauna";
-    }
-
-}
