--- Laborator6/src/SERVICE/Serviciu1.java
@@ -1,23 +1,0 @@
-package SERVICE;
-
-import model1.Animal;
-import model1.exceptii.Exceptie1;
-
-import java.util.List;
-import java.util.Objects;
-
-public class Serviciu1 {
-
-    public List<Animal> animale;
-
-    private void getAnimale(Animal animal){
-        animale.add(animal);
-    }
-
-    public static Boolean utile(List<Object> listaAnimale){
-        if(listaAnimale.isEmpty())
-            throw new Exceptie1("Niciun animal in lista");
-        return true;
-    }
-
-}
