--- Laborator6/src/model1/Animal.java
@@ -1,8 +1,0 @@
-package model1;
-
-public interface Animal {
-
-    public FisaMedicala getFisaMedicala();
-
-    public String vorbeste();
-}
