--- Laborator6/src/model1/Caine.java
@@ -1,13 +1,0 @@
-package model1;
-
-public class Caine {
-    public String NUME;
-
-    public String rasa;
-
-    public String vorbeste(){
-        return "Cainele " + this.NUME + " latra";
-
-    }
-
-}
