--- Laborator6/src/model1/Vaccin.java
@@ -1,12 +1,0 @@
-package model1;
-
-import java.util.Date;
-
-public class Vaccin {
-
-    private String DENUMIRE_VACCIN;
-    public Date data;
-
-
-
-}
