--- Laborator6/src/model1/exceptii/Exceptie1.java
@@ -1,9 +1,0 @@
-package model1.exceptii;
-
-public class Exceptie1 extends RuntimeException{
-
-    public Exceptie1(String message){
-        super(message);
-    }
-
-}
