--- Laborator6/src/model1/exceptii/Exceptie2.java
@@ -1,8 +1,0 @@
-package model1.exceptii;
-
-public class Exceptie2 extends RuntimeException{
-
-    public Exceptie2(String message){
-        super(message);
-    }
-}
