--- Laborator6/src/model1/exceptii/VaccinIntarziatexceptie.java
@@ -1,8 +1,0 @@
-package model1.exceptii;
-
-public class VaccinIntarziatexceptie extends RuntimeException{
-
-    public VaccinIntarziatexceptie(String m){
-        super(m);
-    }
-}
